requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
pycryptodome==3.10.1
pycrypto==2.6.1
gmssl==3.2.1
ecdsa==0.17.0
base58==2.1.0
python-dotenv==0.19.0
Flask-WTF==0.15.1
Flask-Limiter==2.4.0
gunicorn==20.1.0
cryptography==3.4.8
pyOpenSSL==21.0.0
pytest==6.2.5 